refactor(collab): extract server creation and utilities for testability
Extracts createCollabServer() with a typed CollabServerConfig
interface into server.ts, and parseCookies/getDbHostname into
utils.ts. Adds page-access authorization check during auth.

The old index.test.ts (trivial mock test) is removed; proper
tests follow in a subsequent commit.
